// src/plugins/real-estate/ui/generated-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export type Node = {
  id: Scalars['ID'];
};
// ...
export type RealEstate = Node & {
  __typename?: 'RealEstate';
  id: Scalars['ID'];
  projectName: Scalars['String'];
  descriptions: Scalars['String'];
  price: Scalars['String'];
  address: Scalars['String'];
  createdAt: Scalars['DateTime'];
  updatedAt: Scalars['DateTime'];
};

export type RealEstateAddInput = {
  projectName?: Maybe<Scalars['String']>;
  descriptions?: Maybe<Scalars['String']>;
  price?: Maybe<Scalars['String']>;
  address: Scalars['String'];
};
// ...
export type RealEstateUpdateInput = {
  id: Scalars['ID'];
  projectName: Scalars['String'];
  descriptions: Scalars['String'];
  price: Scalars['String'];
  address: Scalars['String'];
};
// ...
export namespace UpdateRealEstate {
  export type Variables = UpdateRealEstateMutationVariables;
  export type Mutation = UpdateRealEstateMutation;
  export type UpdateRealEstate = UpdateRealEstateMutation['updateRealEstate'];
}

export namespace CreateRealEstate {
  export type Variables = CreateRealEstateMutationVariables;
  export type Mutation = CreateRealEstateMutation;
  export type AddRealEstate = CreateRealEstateMutation['addRealEstate'];
}
// ...
export type RealEstateCustomFieldsFragment = (
  { __typename?: 'RealEstate' }
  & Pick<RealEstate, 'id' | 'projectName' | 'descriptions' | 'price' | 'address' | 'createdAt' | 'updatedAt'>
);

export type UpdateRealEstateMutationVariables = Exact<{
  input: RealEstateUpdateInput;
}>;


export type UpdateRealEstateMutation = (
  { __typename?: 'Mutation' }
  & { updateRealEstate: (
    { __typename?: 'RealEstate' }
    & RealEstateCustomFieldsFragment
  ) }
);

export type CreateRealEstateMutationVariables = Exact<{
  input: RealEstateAddInput;
}>;


export type CreateRealEstateMutation = (
  { __typename?: 'Mutation' }
  & { addRealEstate: (
    { __typename?: 'RealEstate' }
    & RealEstateCustomFieldsFragment
  ) }
);